backend/requirements.txt
# rquirements file
fastapi
uvicorn
motor
pydantic-settings
python-dotenv
email-validator
PyJWT